Include changes for custom SinkDSM in new structure
